fix: incorrect asyncapi render
